refactor: edit-cheatsheets画面をコンポーネント/フックへ分割しBadgeを共通化
page.tsx(1512行)をhooksとorganismsに分割し307行へ削減。
TypeBadge/LayoutBadgeのドロップダウン開閉ロジックとPortalシェルを
useBadgeDropdown/BadgeDropdownに共通化。振る舞いは変更なし。
