+++ src/components/organisms/edit-cheatsheets/index.ts
@@ -1,0 +1,3 @@
+export * from './EditRow'
+export * from './LayoutBadge'
+export * from './TypeBadge'
