+++ src/hooks/edit-cheatsheets/constants.ts
@@ -1,0 +1,4 @@
+export const MAX_TITLE = 100
+
+let _uid = 0
+export const nextLocalId = () => `loc${_uid++}`
